Move classes para o pacote entidades
